Modified column update so that unclaimed slabs are skipped.
The issue with drawing Gems was caused by the fact that there
is no gems drawing item in the slabset array. Since gems are
unchangeable, the best solution was to exclude them from update.
To do that, update algorithm was modified to never update
Obstacles and Unclaimable slabs. This means Path will also
no longer be updated when adjacent slab is changed.