Fixed diggers reset on nevverending tasks.
Diggers should be reset when repeatively doing a task which
can't be finished (digging gems) over 10 times in a row.

This functionality was damaged by overcomplicated conditions;
this change fixes the reset on neverending tasks, and adds
a log message when the reset is performed. Some other log
messages were also improved.