Modified computer player traps selling to keep economy.
Selling traps by computer player was configured as a last
resort task when gold amount is critical. It has been
modified to be a standard way of keeping economy running.

Selling for economy upkeep is only allowed to sell
trap and door boxes, and while they are being sold
the computer player will not move creatures from
expensive rooms. The standard path is still triggered
when gold falls too fast, or no traps are left to sell.