Updated search for lair totem to check all subtiles.
When a creature makes lair, it will now check all subtiles
on each slab for free lair space. Still, central subtiles
will be preferred and other ones will be used only if central
are all in use. It is rare for lair to have large enough efficiency
to allow more totems than one per slab, but it happens sometimes.