Added option for 3D Explosion Blast Forces and 3D Splash Damage (#24)
The game is now able tho handle area damage and blast (knockback) in 3D. That means the height of both the exploding projectile and the target creature are considered when calculating the distance between them.

Explosions are now also 3 dimensional, in that they can blast creatures up or down when the explosion is not at their altitude.

How much vertical force the explosions apply can be configured by changing 'ExplosionsVerticalPushPercent' in the rules.cfg. The default is set to 0%, since setting it to a higher value has gameplay impact and may expose pre-existing bugs.