Outro cinematic plays with Lotl in prison
Fixes a bug to the recently introduced drag cinematic. At campaign end it no longer showed the outro cinematic if you won the game with the Knight in prison, showing the drag cinematic instead. Now at the last level it will never do the torture screen or the preceding drag cinematic.