Computer player limits room tiles and sells crates
As part of the "EVENT CHECK ROOMS FULL" computer players will now consider a maximum number of room tiles.
When the full room of maximum size is a workshop, and over half the capacity is in use by crates, the CP will sell a bunch of crates to make space.

Updated level 19 to benefit from this change.

Also, implemented workaround to #135