Use Bresenham’s Line Drawing Algorithm with clipping
draw_stripey_line() has been re-written to use Bresenham’s Line Drawing Algorithm, including clipping the line within the engine_window.

The "line is off the screen" error catches are not required when called from draw_clipped_line(), but are added in case a sanity check is not performed by the calling function (i.e. checking at least some of the line is within the engine window).

A log message will be created if the function tries to write outside of the engine window (this will be removed after more testing). This shouldn't happen, but if it does some of the maths is wrong (likely rounding error).