new swap_assigned_keys() is called when needed
If set_game_key() detects that the currently chosen key configuration is already in use, it assigns that key combo to the current key, and takes the combo from the current key and assigns it to the key that is already using the matching key combo.
i.e. if a maching key combo is found it is swapped with the current key.

One can call  swap_assigned_keys() with the struct GameKey of the matching key combo, and the new key id, key code, and modifier key code to "swap" the key combos.