Pressing Esc to cancel a key assignment no longer returns to the options screen
When the Escape key was being read to cancel a key assignment in the Define Keys menu, the Escape Key value was not cleared.

This meant the Escape key was able to then be passed to the frontscreen_end_input() function, which would then return to the previous menu (the Options Menu).

Now the value is cleared in define_key_input().

Right mouse click has the same behaviour, but this was already handled correctly (the Right Mouse Button value was cleared by define_key_input).