Highlight swapped keys
Any key definition that is swapped, i.e. the user didn't necessarily choose or want to assign this input, is highlighted in grey/dark red (frontend font colour #3).

This happens when a key is assigned a value that was already in use. The "other" key assignment, i.e. the one that takes on the old value of the key currently being assigned, is the one highlighted.