Add ALT as an allowable single modifier key
For Rotate & speed and Possess & query
It can still be used as a combo key for other key assignments.
Single modifier keys can still not be assigned to the "normal" keys (i.e. not Rotate, Speed, Possess, and Query) - this is a limitation of the current implementation. Single keys are currently disallowed, so that it is easier to enter a key combination. That is also why Rotate, Speed, Possess, and Query can only be a single key.