Scrollbar tabs now move when you click and drag
The scrollbars work well with the mouse and mouse wheel in general, but there is one case where it doesn't:
If you click the scrollbar tab and then drag it to a new location, it doesn't visibly move. If you MouseUp, it does jump to the new location, but it does not track with the user's mouse movement.

Changing the scroll_tab menu items from LbBtnT_NormalBtn(0) to LbBtnT_HoldableBtn(1) allows the tabs to move with the mouse.

There is now good visual feedback for this input method.