Add mouse3 and mouse wheel to bindable keys
Middle mouse button (Mouse 3) and mouse wheel up/down can be bound in the key binding menu to an action.

If bound to an action, they action will trigger when the user presses mouse 3, or scrolls the mouse wheel up or down.

The fact the mouse buttons work as key bindings for actions in-game is because of the use of lbKeyOn[].

The use of lbInkey lets the user bind the mouse buttons in the key bind menu.

Uses hardcoded strings for the mouse buttons when shown on the key binding menu, appropriate new ones should probably be added to the dat file.