Use new GUI strings for mouse button binding
The three new strings are used on the Key Binding Menu for mouse wheel up and down, and for numbered mouse buttons.

Mouse 3 (Middle Mouse button) calculates the number 3 from it's position in the enum KeyCodes, relative to it's position to KC_MOUSE1 + 1

As new buttons are added, they can use the same code as just described above, by adding their keycode enums to the switch statement in frontend_draw_define_key().