build/sell advanced controls now use keybinds
One-Click build and Sell Trap on Subtile controls now use is_game_key_pressed() instead of is_key_pressed(). As a result they no longer use hard-coded key bindings, but use the key defined in the keybinding menu.

This input code should be moved to front_input.c, but that will be incorporated in a larger piece of work focused on input.

The three keys Gkey_SellTrapOnSubtile, Gkey_SquareRoomSpace and Gkey_BestRoomSpace are now in an isolated group in the keybind menu, and can be any single key. If CTRL/SHIFT/ALT are used, they can each only be used once in this group (and will not clash with modkey keybinds outside the group).

check_and_assign_mod_keys_group() added to kjm_input.c to allow an isolated group of X keybinds.