Fix log errors related to depleted offmap traps
When a trap was depleted and only offmap traps were available, the code set the health to -1 so it would be destroyed. However, the function that destroyed it increased the available numbers again, causing an inconsistency that was found and corrected in a later check.

Not the trap is deleted directly so no inconsistency is created.