Can now use commented-out lines in .lif files
Start a line with a semi-colon (;) to have it be ignored by the game engine

The next line can be either another commented-out line, or a line describing a level.

semi-colons can still be used in a level name.

You cannot add a comment at the end of a line (it will be treated as part of the level name)