acute fix to #836
This fixes the issue at hand.

However, it only address the acute problem described in that issue.
This fix gives an example of the reorganisation that should happen in the input code (the checking of the scroll wheel performed in maintain_scroll_up/down should likely happen somewhere in front_input.c, and it uses hard-coded controls).