STRIPEY_LINE_COLOR_COUNT moved
STRIPEY_LINE_COLOR_COUNT is now part of the stripey_line_colors enum, rather than being defined as a separate constant.

This means it will automagically count the number of line colours based on the size of the stripey_line_colors enum.
It must remain as the last entry in the stripey_line_colors enum, new colours should be added before it in the list.