Improve One-Click Sell
Adds new "automagic" One-Click Sell - hold SHIFT and hover over a roomslab to highlight the room. Click once to sell all slabs.

"Square Brush" mode now shows the outer box, or brush, in a grey bounding box and the slabs that have valid slabs to sell in a green bounding box.

Fixes 2 bugs where roomspaces of size 1 would not be sold.

Used the new "construction" line colour to indicate that there is currently a roomspace being built/sold - and therefore the player cannot build or sell any rooms. The bounding box primary color switches to this red/yellow colour whilst a build/sale is in progress, and then it returns to normal once the build/sale is complete.

get_current_room_as_roomspace(), and other sub functions, can be used to create a roomspace from a "Room", so that the correct bounding box can be drawn for automagic one-click sell.