new line colour for untagging slabs
A new line colour (currently SLC_YELLOW) is used for the bounding box when the current slab is already tagged for digging. The line will be green when the current slab is not tagged for digging.
When the user holds the left mouse button to paint slabs, then the bounding box will be green when tagging and yellow when untagging