stop pickaxe cursor being stuck on
if the user holds LMB on a dungeon message, and the hovers over a slab, they got locked in dungeonadd->one_click_lock_cursor mode until they click either mouse button.

This is now fixed.

This happened because player->cursor_button_down = 0 in this situation, and the new code for one-click had assumed player->cursor_button_down would always be 1 when LMB was held.