position tool tips correctly
Tool Tips are positioned according to the mouse X and Y coordinate. A y-offset of 86 was added, which was correct for 640x400.

This fixed 86 has now been replaced with a value based on the units_per_pixel value. This means that the tool tip is now in the correct position at all resolutions.