show hostOS cursor on focus lost
when the keeperFX window has lost focus, show the (e.g.) Windows cursor over the keeperFX window, as it is now "in the background". The DK cursor is restored on focus gain (and the current mouse position is used if the user clicks in the window to re-enter focus).

altinput need to be improved in another commit: it does not disable the kfx cursor on lose focus.