altinput: synchronise Host OS cursor pos and game cursor pos
on loading a screen/level etc

use `LbMouseSetPositionInitial` when initialising the cursor position to the centre of the window, use the old `LbMouseSetPosition` for other purposes (i.e. when we could be in altinput mode and we don't want to move the Host OS cursor too, or we know we are not in altinput mode).

`LbMouseSetPositionInitial` is the same as `LbMouseSetPosition` except in altinput mode where it will first move the game cursor to the position of the Host OS cursor, and then move the Host OS cursor to the given x and y location.