bugfix: stop cursor lock on level start
if the game switches res between menu/splash screen and in-game - this causes the window to gain focus whilst zooming out of the heart. This was grabbing the cursor (if LOCK_CURSOR_IN_POSSESSION=ON) because this is possession mode. GNFldD_Unkn08 of game.numfield_D is now checked, as this seems to be 1 when in this special mode. Checking that GNFldD_Unkn08  is 0 stops the curosr being locked in this edge case.