Delete non-functional continue-game files to make continue save work again (#2151)
When a new version of KeeperFX has changed the 'game' structure, the 'continue' saves will fail to load. However, this now outdated continue-save file would prevent new and correct continue-saves from ever being created, so now it's deleted so the function works again.

This does not solve the continue-saves breaking on updates in the first place.