Timebomb crspell icon updated
Incredible minor and petty update to the icon of a spell nobody uses - Timebomb creature spell points to the "creature age" icon in the files, which is an odd size and changes to the skill icon once used. I had previously made a better fitting version of this icon with an associated dis version, and now that's in this can point to it.