Made vampire on lvl9 white again
reverted the dragonslover change making the white vampire neutral on moonbrush wood.

To compensate, I checked if the player entered the center and if not the heroes drop in from one of the question marks.

Finally, made sure the 'vampire from graveyard' message only spawned when you get one from your graveyard.