allow choosing of display via config file
Default (with no value in cfg file) is 0, which is the primary/first/only display.

Setting `display_number=0` is exactly the same as current master (`SDL_WINDOWPOS_CENTERED` is the same as `SDL_WINDOWPOS_CENTERED_DISPLAY(0)`)

Usage:
e.g. putting the following in the config file will cause keeperfx to run on the 2nd monitor.
```
; The display to use for the game
DISPLAY_NUMBER=1
```
implements #2345