Allow CPU to bridge over water when not wall hugging
We do not want to run `find_from_task_list` before queueing a bridge, because this is the imp's task list, and bridging is not an imp task so won't be listed.

When handling water slabs, we want to leave `tool_dig_to_pos2_skip_slabs_which_dont_need_digging_f` when we have bridges available, otherwise we want to treat the water as passable.

When handling lava slabs, we want to leave `tool_dig_to_pos2_skip_slabs_which_dont_need_digging_f` always, because we either want to bridge it, or it is impassable and should be wallhugged.