Computer Players no longer flawless in job assignment (#3052)
`MOVE CREATURE TO BEST ROOM` and `MOVE CREATURE TO ROOM` computer processes accept a percentage of creatures that would be assigned to a good room. But it would move that percentage each check, so it quickly had 100% of creatures in a good room. Now it will stick to that percentage by checking how many units are already assigned before moving more creatures.

By default, they are nearly flawless still. Opens up more limited computer players in a future version.