Preload all sound samples
This removes the sound heap which was used as a cache to store samples.
Samples were loaded on-demand directly from disk as a means to save memory.

Loading sound samples like this is at least partially responsible
for the stutter every time a game is started or loaded.

This is one of the steps needed for OpenAL conversion.