Fix Heal in the Possession mode
The Heal spell is broken in Possession mode because the recent
chagne added a shot model for it.
Currently, the Heal instance in the Possession mode has no target
index specified, so the Heal will be fired as a shot.

This change improves get_human_controlled_creature_target() to
ensure it can find the target for self-buff spells.
It can incorporates the new feature made in 7ab2ff6c.

Type: Bug Fix