Fix crash when searching creatures for ranged buff
The program will crash when there are 3+ creatures being included
in the result list of search_target_generic().
The reason is unclear. It seems that C has special rules for double
star (**) type. It will corrupt the memory when access *pp[i] if
i >= 2 even if *pp was created by malloc() with enough space.

For exmaple:
void fooint( int **pp )
{
  *pp = malloc(100);
  *pp[2] = 9; //crash.
  // It must be rewritten in:
  int *p2 = malloc(100);
  p2[2] = 9; // ok.
  *pp = p2;
}

Type: Bug Fix