For ranged buffs except heal, forbid targeting imprisoned allies.
If our creatures are imprisoned, it's not reasonable to cast buffs
on them because they are not valid to fight.
But heal is allowed because imprisoned creatures will starve to death
unless they get healed.
If our creatures break enemy's defense and reach the prison, we should
allow our ranged healers to save their allies from death.

Type: New Feature