Ranged buff: don't fire at void in Possession mode
Currently, the ranged buffs are allowed to fire at nothing.
A SHOT will be fired and it may hit something that we don't want,
for example, unconscious creatures.

This is a waste. The change prevents it. Now unless the player aim
at a valid creature, the ranged buff won't be used.

This change won't affect debuffs or ranged attacks.

Type: Bug Fix