Set default case to -3
Made it so the synergy slab is set to -3 if the parameter cannot be read, it should be more robust than leaving it unset since slab0 is actually used in game. Furthermore that value is different from the -2 it gets in case the slab was set to none, which bugs more traceable.