Various bug fixes
The log highlighted issues with invalid control being tested. To fix this, ensure creature view icons are not drawn if control is invalid. Control can become invalid if the creature dies and its structure is deleted, but the player doesn't leave possession. Additionally, check for invalid control in `compute_creature_weight`. Only compute if control is valid, and return 0 otherwise.

For `delete_thing_structure_f`, instead of checking the class of the thing, check for valid control for creature-related actions.

Lastly, fix the loop in `draw_creature_view_icons` to only check for active spells. I encountered a bug where spells that were not cast appeared in the possession view because they shared the same spell flags as an active spell.