Added scavenge success text string and icon, started work updating heart attacked message
Also updated room hotkey names.
Considered also updating gui_area_event_button in frontmenu_ingame_tabs.c so that the heart attack message displays the tab in yellow when you click it instead of red (like the combat button), but I'm not sure what would be easier here: making a specific case for it where it links to 260 when highligted and 777 when not, or reordering all of the menu sprites so that there is a copy of message_rpanel_msg_exclam_act at 777, message_rpanel_msg_exclam2_act is 778, and everything else is moved down (then I change the offset to +2).