Fixed dropped units counted as enemy controlled (#4057)
Units that you picked up from an enemy room and you dropped in your dungeon but not on a room, would temporarily be counted as being controlled by the enemy while they were in the air.
This because the game assumed they were being dropped by the enemy on a torture room or prison.