some sort of attempt to add manufacturable icon
Not sure if any of this makes sense so please check.

Aim is: if icon appears (i.e. you have or had the trap/door), display hammer icon in top-right of panel (i.e. over frame_wide_empty): if buildable and player's manufacture level is high enough - icon will be lit up. If not high enough level, it will be greyed out. If it is not manufacturable, an icon for "can't manufacture" will be shown.

Do we want this to also appear even if the icon is a ?, might need more stuff changed if so.

I'm not sure if I've linked up set_manufacture_level and is_trap/door_buildable so let me know.