That has offset between the calculate character and the draw character.
When word are very long, they will reproduce.

The issues caused by offset in multi line text are as follows.
The character that should have been displayed at the end of the first line is pushed to the next line.
And this may lead to subsequent text being cut off.